+++ crates/tauri/src/path/mod.rs
@@ -124,7 +124,7 @@ pub enum BaseDirectory {
   /// Resolves to [`crate::path::PathResolver::video_dir`].
   Video = 10,
   /// The Resource directory.
-  /// Resolves to the resource directory of this app.
+  /// Resolves to [`crate::path::PathResolver::resource_dir`].
   Resource = 11,
   /// A temporary directory.
   /// Resolves to [`std::env::temp_dir`].